pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "photokinthesis"
version = "0.1.0"
description = "A Python package"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "typer[all]>=0.9.0",
    "imagehash>=4.3.0",
    "Pillow>=10.0.0",
    "mediapipe>=0.10.0",
    "opencv-python>=4.5.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "mtcnn>=0.1.1",
    "tensorflow>=2.13.0",
]

[project.scripts]
photokinthesis = "photokinthesis.cli:app"

[tool.uv]
dev-dependencies = []
